# packages/server/api/menu_api/menu_creat.go
package menu_api

import (
	"web/global"
	"web/models"
	"web/models/ctype"
	"web/models/res"

	"github.com/gin-gonic/gin"
)


type ImageSort struct {
  ImageID uint `json:"image_id"`
  Sort    int  `json:"sort"`
}

type MenuRequest struct {
  Title         string      `json:"title" binding:"required" msg:"请完善菜单名称" structs:"title"`
  Path          string      `json:"path" binding:"required" msg:"请完善菜单路径" structs:"path"`
  Slogan        string      `json:"slogan" structs:"slogan"`
  Abstract      ctype.Array `json:"abstract" structs:"abstract"`
  AbstractTime  int         `json:"abstract_time" structs:"abstract_time"`                // 切换的时间，单位秒
  BannerTime    int         `json:"banner_time" structs:"banner_time"`                    // 切换的时间，单位秒
  Sort          int         `json:"sort" binding:"required" msg:"请输入菜单序号" structs:"sort"` // 菜单的序号
  ImageSortList []ImageSort `json:"image_sort_list" structs:"-"`                          // 具体图片的顺序
}

// @Tags 菜单管理
// @Summary 菜单创建
// @Description 传入参数创建菜单
// @Param cr body MenuRequest true "添加的菜单信息"
// @Router /api/menus [post]
// @Accept json
// @Produce json
// @Success 200 {object} res.Response
func (MenuApi) MenuCreateView(c *gin.Context) {
  var cr MenuRequest
  err := c.ShouldBindJSON(&cr)
  if err != nil {
    res.FailWithError(err, &cr, c)
    return
  }

  // 重复值判断
  var menuList []models.MenuModel
  count := global.DB.Find(&menuList, "title = ? or path = ?", cr.Title, cr.Path).RowsAffected
  if count > 0 {
    res.FailWithMessage("重复的菜单", c)
    return
  }

  // 创建banner数据入库
  menuModel := models.MenuModel{
    Title:        cr.Title,
    Path:         cr.Path,
    Slogan:       cr.Slogan,
    Abstract:     cr.Abstract,
    AbstractTime: cr.AbstractTime,
    BannerTime:   cr.BannerTime,
    Sort:         cr.Sort,
  }

  err = global.DB.Create(&menuModel).Error

  if err != nil {
    global.Log.Error(err)
    res.FailWithMessage("菜单添加失败", c)
    return
  }
  if len(cr.ImageSortList) == 0 {
    res.OkWithMessage("菜单添加成功", c)
    return
  }

  var menuBannerList []models.MenuBannerModel

  for _, sort := range cr.ImageSortList {
    // 这里也得判断image_id是否真正有这张图片
    menuBannerList = append(menuBannerList, models.MenuBannerModel{
      MenuID:   menuModel.ID,
      BannerID: sort.ImageID,
      Sort:     sort.Sort,
    })
  }
  // 给第三张表入库
  err = global.DB.Create(&menuBannerList).Error
  if err != nil {
    global.Log.Error(err)
    res.FailWithMessage("菜单图片关联失败", c)
    return
  }
  res.OkWithMessage("菜单添加成功", c)
}